initial implementation of rarity coloring
