+++ enclone_core/Cargo.toml
@@ -16,6 +16,7 @@ publish = false
 # in the root of the enclone repo.
 
 [dependencies]
+amino = "0.1.2"
 ansi_escape = "0.1.0"
 bio = ">=0.31, <=0.32"
 bytes = "0.5.5"
@@ -28,6 +29,8 @@ serde = "1.0.90"
 serde_derive = "1.0.102"
 serde_json = "*"
 string_utils = "0.1.1"
+superslice = "1"
+vdj_ann = { git = "https://github.com/10XGenomics/rust-toolbox.git", rev="fe14ff4f5a6d33c8c28b753c8659f9ca08df9f78" }
 vector_utils = "0.1.3"
 
 [dependencies.hdf5]
