+++ enclone_core/src/defs.rs
@@ -422,6 +422,7 @@ pub struct GeneralOpt {
     pub tree: String,
     pub allow_inconsistent: bool,
     pub color: String,
+    pub color_by_rarity_pc: f64,
     pub species: String, // human or mouse or unknown, determined from the reference sequence
     pub using_secmem: bool,
     pub const_igh: Option<Regex>,
