+++ enclone_print/Cargo.toml
@@ -15,6 +15,8 @@ publish = false
 # Please do not edit crate versions within this file.  Instead edit the file master.toml
 # in the root of the enclone repo.
 
+# To keep compile time down, this file should *not* access the enclone crate.
+
 [dependencies]
 amino = "0.1.2"
 ansi_escape = "0.1.0"
