+++ enclone_tail/Cargo.toml
@@ -15,6 +15,8 @@ publish = false
 # Please do not edit crate versions within this file.  Instead edit the file master.toml
 # in the root of the enclone repo.
 
+# To keep compile time down, this file should *not* access the enclone crate.
+
 [dependencies]
 amino = "0.1.2"
 ansi_escape = "0.1.0"
@@ -34,7 +36,6 @@ serde_derive = "1.0.102"
 serde_json = "*"
 stats_utils = "0.1.1"
 string_utils = "0.1.1"
-superslice = "1"
 tables = "0.1.2"
 tar = "0.4.29"
 vdj_ann = { git = "https://github.com/10XGenomics/rust-toolbox.git", rev="fe14ff4f5a6d33c8c28b753c8659f9ca08df9f78" }
